AGS: Adding shared/gui/ folder

+++ engines/ags/shared/debugging/out.h
@@ -132,19 +132,19 @@ struct DebugGroupID {
 	uint32_t    ID;
 	String      SID;
 
-	DebugGroupID() : ID(kDbgGroup_None) {
+	DebugGroupID() : ID((uint32_t)kDbgGroup_None) {
 	}
 	DebugGroupID(uint32_t id, const String &sid = "") : ID(id), SID(sid) {
 	}
-	DebugGroupID(const String &sid) : ID(kDbgGroup_None), SID(sid) {
+	DebugGroupID(const String &sid) : ID((uint32_t)kDbgGroup_None), SID(sid) {
 	}
 	// Tells if any of the id components is valid
 	bool IsValid() const {
-		return ID != kDbgGroup_None || !SID.IsEmpty();
+		return ID != (uint32_t)kDbgGroup_None || !SID.IsEmpty();
 	}
 	// Tells if both id components are properly set
 	bool IsComplete() const {
-		return ID != kDbgGroup_None && !SID.IsEmpty();
+		return ID != (uint32_t)kDbgGroup_None && !SID.IsEmpty();
 	}
 };
